Move build spi to spi.builder
