+++ java/execute/src/com/ibm/streamsx/topology/spi/operators/Source.java
@@ -7,7 +7,7 @@ import com.ibm.streams.operator.model.SharedLoader;
 import com.ibm.streamsx.topology.function.FunctionContext;
 import com.ibm.streamsx.topology.function.Supplier;
 import com.ibm.streamsx.topology.internal.functional.ops.FunctionSource;
-import com.ibm.streamsx.topology.spi.Invoker;
+import com.ibm.streamsx.topology.spi.builder.Invoker;
 import com.ibm.streamsx.topology.spi.runtime.FunctionalOperator;
 
 /**
