+++ java/src/com/ibm/streamsx/topology/internal/core/SourceInfo.java
@@ -5,7 +5,7 @@
 package com.ibm.streamsx.topology.internal.core;
 
 import static com.ibm.streamsx.topology.internal.gson.GsonUtilities.objectCreate;
-import static com.ibm.streamsx.topology.spi.SourceInfo.SOURCE_LOCATIONS;
+import static com.ibm.streamsx.topology.spi.builder.SourceInfo.SOURCE_LOCATIONS;
 
 import com.google.gson.JsonObject;
 import com.ibm.streamsx.topology.builder.BOperatorInvocation;
@@ -39,7 +39,7 @@ public class SourceInfo {
                 
         JsonObject holder = objectCreate(bop._json(), SOURCE_LOCATIONS);
         
-        com.ibm.streamsx.topology.spi.SourceInfo.addSourceInfo(holder, calledClass);    
+        com.ibm.streamsx.topology.spi.builder.SourceInfo.addSourceInfo(holder, calledClass);    
     }
     
     public static void setSourceInfo(BOperatorInvocation bop, JsonObject config) {
